Compatibility tweak for 64 bit Windows OpenPhone
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@30853 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ samples/openphone/main.cxx
@@ -5350,7 +5350,7 @@ bool OptionsDialog::TransferDataFromWindow()
     PwxString alias = m_Aliases->GetString(i);
     m_manager.h323EP->AddAliasName(alias);
     wxString key;
-    key.sprintf(wxT("%u"), i+1);
+    key.sprintf(wxT("%u"), (unsigned)i+1);
     config->Write(key, alias);
   }
